+++ packages/twenty-front/src/modules/object-record/record-board/states/isSwitchingToKanbanViewTypeComponentState.ts
@@ -1,9 +1,9 @@
 import { ContextStoreComponentInstanceContext } from '@/context-store/states/contexts/ContextStoreComponentInstanceContext';
 import { createComponentState } from '@/ui/utilities/state/component-state/utils/createComponentState';
 
-export const switchesToKanbanViewTypeComponentState =
+export const isSwitchingToKanbanViewTypeComponentState =
   createComponentState<boolean>({
-    key: 'switchesToKanbanViewTypeComponentState',
+    key: 'isSwitchingToKanbanViewTypeComponentState',
     componentInstanceContext: ContextStoreComponentInstanceContext,
     defaultValue: false,
   });
